can move up and down and turn CHECKPOINT but not turn to the right positions

+++ ros2_ws/src/claybot/include/control_turn.hpp
@@ -33,6 +33,6 @@ private:
     int present_position;
 };
 
-uint32_t degreesToPosition(float degrees);
+uint32_t degreesToPosition(int degrees);
 
 #endif  // CONTROL_TURN_HPP_
